Found Bug
After creating a Set of movelists (represented as Arrays) I tried checking to see if that Set.has a particular array that I knew it did but got "false" anyway because Arrays are Objects.  To Resolve this I  am going back to basics and converting my basic moveList state variable from an array to a string and working with strings as much as possible.